+++ src/FlaUILibrary/flaui/module/property.py
@@ -41,6 +41,7 @@ class Property(ModuleInterface):
         IS_TEXT_PATTERN_SUPPORTED = "IS_TEXT_PATTERN_SUPPORTED"
         IS_TOGGLE_PATTERN_SUPPORTED = "IS_TOGGLE_PATTERN_SUPPORTED"
         IS_VALUE_PATTERN_SUPPORTED = "IS_VALUE_PATTERN_SUPPORTED"
+        VALUE = "VALUE"
         IS_EXPAND_COLLAPSE_PATTERN_SUPPORTED = "IS_EXPAND_COLLAPSE_PATTERN_SUPPORTED"
         EXPAND_COLLAPSE_STATE = "EXPAND_COLLAPSE_STATE"
 
@@ -160,6 +161,7 @@ class Property(ModuleInterface):
             self.Action.IS_TEXT_PATTERN_SUPPORTED: lambda: self._is_text_pattern_supported(values["element"]),
             self.Action.IS_TOGGLE_PATTERN_SUPPORTED: lambda: self._is_toggle_pattern_supported(values["element"]),
             self.Action.IS_VALUE_PATTERN_SUPPORTED: lambda: self._is_value_pattern_supported(values["element"]),
+            self.Action.VALUE: lambda: self._get_value_pattern_value(values["element"]),
             self.Action.IS_EXPAND_COLLAPSE_PATTERN_SUPPORTED: lambda: self._is_expand_collapse_pattern_supported(values["element"]),
             self.Action.EXPAND_COLLAPSE_STATE: lambda: self._get_expand_collapse_pattern_state(values["element"]),
         }
@@ -340,3 +342,16 @@ class Property(ModuleInterface):
         pattern = Property._get_expand_collapse_pattern_from_element(element)
         return str(pattern.ExpandCollapseState)
 
+    @staticmethod
+    def _get_value_pattern_from_element(element) -> Any:
+        if Property._is_value_pattern_supported(element):
+            pattern = element.Patterns.Value.Pattern
+            if pattern is not None:
+                return pattern
+
+        raise FlaUiError(FlaUiError.PatternNotSupported.format("Value"))
+
+    @staticmethod
+    def _get_value_pattern_value(element: Any) -> str:
+        pattern = Property._get_value_pattern_from_element(element)
+        return str(pattern.Value)